Actor test (#4)
adds tests to actor package

+++ src/watchableMessageStore.js
@@ -18,11 +18,24 @@
 
 
 import { _throw } from '@dxworks/utils';
-
+// This interval needs to be strictly longer than the time it takes to paint
+// 1 frame. E.g. this value needs to be higher than 16ms. Otherwise, the
+// IDB connection will starve and run into an endless loop.
 const POLLING_INTERVAL = 50;
 const DB_PREFIX = 'ACTOR-DATABASE';
 const OBJECT_STORE_NAME = 'LIST';
-
+/**
+ * A messageStore that can read and write to a specific objectStore in an
+ * IndexedDB database. This class is used to implement message passing for
+ * actors and is generally not intended to be used for other use cases.
+ *
+ * To retrieve messages from the store for a recipient, use
+ * {@link watchableMessageStore~popMessages}. To subscribe to newly added
+ * messages for a recipient, use {@link watchableMessageStore~subscribe}.
+ * To store a new message for a recipient, use
+ * {@link watchableMessageStore~pushMessage}.
+ * @param {string} _name
+ */
 export const watchableMessageStore = _name => {
   const name = _name;
   const objStoreName = OBJECT_STORE_NAME;
@@ -54,8 +67,16 @@ export const watchableMessageStore = _name => {
   const database = init();
 
   const bcc = ('BroadcastChannel' in self) && new BroadcastChannel(name);
-
-  const popMessages = async (recipient, { keepMessage = false } = {}) => {
+  /**
+   * Retrieve all messages for a specific recipient. You can specify with
+   * `keepMessage` whether to keep messages after they are processed by
+   * the recipient.
+   *
+   * @param {string} recipient The name of the recipient.
+   * @param {boolean} [keepMessage=false] Whether to keep any messages after the recipient has
+   *    processed the message.
+   */
+  const popMessages = async (recipient, keepMessage = false) => {
     const transaction = (await database).transaction(objStoreName, 'readwrite');
     const cursorRequest = transaction
       .objectStore(objStoreName)
@@ -67,23 +88,28 @@ export const watchableMessageStore = _name => {
       };
       cursorRequest.onsuccess = () => {
         const cursor = cursorRequest.result;
-        if (cursor) {
+        const addToMessages = value => {
+          messages.push(value);
+          (!keepMessage) && cursor.delete();
+        };
+        const updateLastCursorId = () => {
           const { value } = cursor;
-          if (value.recipient === recipient || recipient === '*') {
-            messages.push(value);
-            if (!keepMessage) {
-              cursor.delete();
-            }
-          }
+          (value.recipient === recipient || recipient === '*') &&
+            addToMessages(value);
           cursor.continue();
           lastCursorId = cursor.key;
-        } else {
-          resolve(messages);
-        }
+        };
+        cursor
+        ? updateLastCursorId()
+        : resolve(messages);
       };
     });
   };
-
+  /**
+   * Store a message with a recipient.
+   *
+   * @param {{recipient: string, handler: string , detail}} message The message to store with a recipient and a detail.
+   */
   const pushMessage = async message => {
     (message.recipient === '*') && _throw('Can’t send a message to reserved name "*"');
     const transaction = (await database).transaction(objStoreName, 'readwrite');
@@ -98,17 +124,19 @@ export const watchableMessageStore = _name => {
       transaction.objectStore(objStoreName).add(message);
     });
   };
+  /**
+   * @param {string} recipient The name of the recipient.
+   * @param {function} callback The callback that is invoked with all new messages.
+  */
   const subscribeWithBroadcastChannel = (recipient, callback) => {
     const channel = new BroadcastChannel(name);
+    const retrieveRecipientMessages = async () => {
+      const messages = await popMessages(recipient);
+      (messages.length > 0) && callback(messages);
+    };
     const channelCallback = async evt => {
       const ping = evt.data;
-      if (ping.recipient !== recipient) {
-        return;
-      }
-      const messages = await popMessages(recipient);
-      if (messages.length > 0) {
-        callback(messages);
-      }
+      (ping.recipient === recipient) && await retrieveRecipientMessages();
     };
     channel.addEventListener('message', channelCallback);
     // Check for already stored messages immediately
@@ -117,13 +145,15 @@ export const watchableMessageStore = _name => {
       channel.close();
     };
   };
+  /**
+   * @param {string} recipient The name of the recipient.
+   * @param {function} callback The callback that is invoked with all new messages.
+  */
   const subscribeWithPolling = (recipient, callback) => {
     let timeout = -1;
     const pollCallback = async () => {
       const messages = await popMessages(recipient);
-      if (messages.length > 0) {
-        callback(messages);
-      }
+      (messages.length > 0) && callback(messages);
       timeout = setTimeout(pollCallback, POLLING_INTERVAL);
     };
     timeout = setTimeout(pollCallback, POLLING_INTERVAL);
@@ -131,13 +161,21 @@ export const watchableMessageStore = _name => {
       self.clearTimeout(timeout);
     };
   };
-
+  /**
+   * Add a callback whenever a new message arrives for the recipient.
+   * Depending on the functionality of your browser, this will either use
+   * `BroadcastChannel` for fine-grained notification timing. If the browser
+   * does not implement `BroadcastChannel`, it falls back to use a polling
+   * mechanism. The polling timeout is specified by {@link POLLING_INTERVAL}.
+   * @param {string} recipient The name of the recipient.
+   * @param {function} callback The callback that is invoked with all new messages.
+  */
   const subscribe = (recipient, callback) => (
     ('BroadcastChannel' in self)
       ? subscribeWithBroadcastChannel(recipient, callback)
       : subscribeWithPolling(recipient, callback)
   );
-  return message => (
+  return /** @param {string} message */ message => (
     message === 'resetCursor'
     ? resetCursor
     : message === 'init'
